docker/llm-service/app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.7
requests==2.31.0
faiss-cpu==1.7.4
numpy==1.26.0
python-multipart==0.0.6
pypdf==3.12.0
sentence-transformers==2.2.2
pytest==7.4.0
python-jose==3.3.0
